+++ index.js
@@ -2,13 +2,15 @@ var fs = require('fs')
 var path = require('path')
 var mkdirp = require('mkdirp')
 var includePathSearcher = require('include-path-searcher')
-var quickTemp = require('quick-temp')
+var Writer = require('broccoli-writer')
 var mapSeries = require('promise-map-series')
 var sass = require('node-sass')
 var _ = require('lodash')
 
 
 module.exports = SassCompiler
+SassCompiler.prototype = Object.create(Writer.prototype)
+SassCompiler.prototype.constructor = SassCompiler
 function SassCompiler (sourceTrees, inputFile, outputFile, options) {
   if (!(this instanceof SassCompiler)) return new SassCompiler(sourceTrees, inputFile, outputFile, options)
   this.sourceTrees = sourceTrees
@@ -23,10 +25,9 @@ function SassCompiler (sourceTrees, inputFile, outputFile, options) {
   }
 }
 
-SassCompiler.prototype.read = function (readTree) {
+SassCompiler.prototype.write = function (readTree, destDir) {
   var self = this
-  quickTemp.makeOrRemake(this, '_tmpDestDir')
-  var destFile = this._tmpDestDir + '/' + this.outputFile
+  var destFile = destDir + '/' + this.outputFile
   mkdirp.sync(path.dirname(destFile))
   return mapSeries(this.sourceTrees, readTree)
     .then(function (includePaths) {
@@ -37,10 +38,5 @@ SassCompiler.prototype.read = function (readTree) {
       _.merge(sassOptions, self.sassOptions)
       var css = sass.renderSync(sassOptions)
       fs.writeFileSync(destFile, css, { encoding: 'utf8' })
-      return self._tmpDestDir
     })
-}
-
-SassCompiler.prototype.cleanup = function () {
-  quickTemp.remove(this, '_tmpDestDir')
 }